Customize JSON parsing of configuration
Since the default implementation will overwrite the value of duplicate
keys a custom check function is introduced.
This function will raise a KeyError upon any duplicate key in the parsed
JSON.

The reason for this is to prohibit duplicate configuration keys as it
might confuse people when maintaining large configurations.

This fixes #1