Introduce kubernetes context awareness
Make ausroller check for the supplied kubernetes context to be
available.
This checks if kubectl is has configured the context passed via cmdline
and uses the context when actually applying resources.

The repopath configured via the config file is now also defined per
context. So one can define multiple repositories and use them via the
context switch (-C | --context).
This way you can not accidentally commit to the wrong repository.

A change to the configuration file (ausroller.ini) is needed for this to
work. Now one has to provide a section per context to use e.g.:

[<k8s_context_name>]
repopath = <path/to/repo>

close #4